+++ packages/table/package.json
@@ -17,9 +17,9 @@
     "scripts": {
         "clean": "rm -rf dist/* || rm -rf lib/* || true",
         "compile": "run-p \"compile:*\"",
-        "compile:esm": "tsc -p ./src",
-        "compile:cjs": "tsc -p ./src -m commonjs --verbatimModuleSyntax false --outDir lib/cjs",
-        "compile:esnext": "tsc -p ./src -t esnext --outDir lib/esnext",
+        "compile:esm": "tsc -p ./src/tsconfig.build.json",
+        "compile:cjs": "tsc -p ./src/tsconfig.build.json -m commonjs --verbatimModuleSyntax false --outDir lib/cjs",
+        "compile:esnext": "tsc -p ./src/tsconfig.build.json -t esnext --outDir lib/esnext",
         "compile:css": "sass-compile ./src",
         "dev": "run-p \"compile:esm --watch\" \"compile:css --watch\"",
         "dist": "run-s \"dist:*\"",
@@ -30,8 +30,8 @@
         "lint:es": "es-lint",
         "lint-fix": "es-lint --fix && sass-lint --fix",
         "test": "run-s test:typeCheck test:iso test:karma",
-        "test:typeCheck": "tsc -p ./test",
-        "test:iso": "mocha test/isotest.mjs",
+        "test:typeCheck": "tsc -p ./src/tsconfig.test.json",
+        "test:iso": "mocha src/isotest.mjs",
         "test:karma": "karma start",
         "test:karma:debug": "karma start --single-run=false --reporters=helpful --debug",
         "verify": "npm-run-all compile -p dist test lint"
