+++ packages/table/src/tsconfig.build.json
@@ -2,7 +2,6 @@
     "extends": "../../../config/tsconfig.web",
     "compilerOptions": {
         "composite": true,
-        "lib": ["dom", "es5", "es6"],
         "outDir": "../lib/esm"
     },
     "exclude": [
