+++ packages/table/test/index.ts
@@ -16,32 +16,36 @@
 
 import "@blueprintjs/test-commons/bootstrap";
 
-import "./batcherTests";
-import "./cellTests";
-import "./clipboardTests";
-import "./columnHeaderCellTests";
-import "./columnTests";
-import "./common/internal";
-import "./editableCellTests";
-import "./editableNameTests";
-import "./formats/jsonFormatTests";
-import "./formats/truncatedFormatTests";
-import "./gridTests";
-import "./guidesTests";
-import "./loadableContentTests";
-import "./loadingOptionsTests";
-import "./locatorTests";
-import "./menusTests";
-import "./quadrants/tableQuadrantStackTests";
-import "./quadrants/tableQuadrantTests";
-import "./rectTests";
-import "./regionsTests";
-import "./reorderableTests";
-import "./resizableTests";
-import "./rowHeaderCellTests";
-import "./selectableTests";
-import "./selectionTests";
-import "./tableTests";
-import "./tableBodyTests";
-import "./utilsTests";
-import "./resizeRowsTests";
+import "../src/common/batcher.test";
+import "../src/cell/cell.test";
+import "../src/common/clipboard.test";
+import "../src/headers/columnHeaderCell.test";
+import "../src/column.test";
+import "../src/common/internal/directionUtils.test";
+import "../src/common/internal/focusedCellUtils.test";
+import "../src/common/internal/platformUtils.test";
+import "../src/common/internal/scrollUtils.test";
+import "../src/common/internal/selectionUtils.test";
+import "../src/cell/editableCell.test";
+import "../src/headers/editableName.test";
+import "../src/cell/formats/jsonFormat.test";
+import "../src/cell/formats/truncatedFormat.test";
+import "../src/common/grid.test";
+import "../src/layers/guides.test";
+import "../src/common/loadableContent.test";
+import "../src/loadingOptions.test";
+import "../src/locator.test";
+import "../src/interactions/menus/menus.test";
+import "../src/quadrants/tableQuadrantStack.test";
+import "../src/quadrants/tableQuadrant.test";
+import "../src/common/rect.test";
+import "../src/regions.test";
+import "../src/interactions/reorderable.test";
+import "../src/interactions/resizable.test";
+import "../src/headers/rowHeaderCell.test";
+import "../src/interactions/selectable.test";
+import "../src/selection.test";
+import "../src/table.test";
+import "../src/tableBody.test";
+import "../src/common/utils.test";
+import "../src/resizeRows.test";
